Moves AssignWidget to queue/components
